+++ pages/cards/[id].tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { useRouter } from "next/router";
 import Heading from "@/components/heading/header";
 import Nav from "@/components/nav";
 import PathBar from "@/components/pathBar";
@@ -7,7 +6,6 @@ import styled from "styled-components";
 import Footer from "@/components/footer";
 import Image from "next/image";
 import { BsChat } from "react-icons/bs";
-
 import { getPostDetails, getPostIdList } from "../../lib/posts";
 import BottomNav from "@/components/bottomNav";
 import PhotoBox from "@/components/photoBox";
@@ -15,31 +13,8 @@ import QuantityBox from "@/components/quantityBox";
 import SpecificDetails from "@/components/specificDetails";
 import { CardInfo } from "@/types";
 import { InferGetStaticPropsType } from "next";
-import { WishListItem } from "@/store/wishListReducer";
-import { useUser } from "@/store/userReducer";
 
 const CardDetail = ({ cardData }: { cardData: CardInfo }) => {
-  const router = useRouter();
-  const { userObj } = useUser();
-  console.log({ userObj });
-  const userId = userObj ? userObj.id : "";
-  console.log(userId);
-
-  const { id } = router.query;
-  const itemId = id ? parseInt(id as string, 10) : 0;
-
-  const item: WishListItem = {
-    itemId: itemId,
-    name: cardData.infoTitle,
-    description: cardData.infoDetail,
-    userId,
-    mainImage: cardData.mainImage,
-    profileId: cardData.profileId,
-    price: cardData.price,
-  };
-  console.log(item);
-  console.log(cardData);
-
   return (
     <BgWrapper>
       <Main>
@@ -77,7 +52,7 @@ const CardDetail = ({ cardData }: { cardData: CardInfo }) => {
         <QuantityBox />
         <SpecificDetails cardData={cardData} />
         <Footer />
-        <BottomNav item={item} cardData={cardData} />
+        <BottomNav cardData={cardData} />
       </Main>
     </BgWrapper>
   );
